% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factRedux-package.R, R/factRedux.R
\docType{package}
\name{factRedux}
\alias{factRedux}
\alias{factRedux-package}
\title{VasRoi: A package for analysis of return on investment from Value Added Services}
\usage{
factRedux(data, factor = "", levels = 10, inc.other = TRUE)
}
\arguments{
\item{factor}{factor to be simplified}
}
\description{
The driversOfPerf package provides three categories of functions: Function for data
processing, function for impacts analysis and function for plotting

factRedux Function
}
\details{
Function to create an other category in functions where there are too many different levels to present cleanly
}
\author{
Ben Downe
}

