% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factRedux.R
\name{factRedux}
\alias{factRedux}
\title{factRedux Function}
\usage{
factRedux(data, factor = "", new.level = "Other", levels = 10,
  inc.other = TRUE)
}
\arguments{
\item{factor}{factor to be simplified}

\item{new.level}{factor level to replace levels outside those being considered}

\item{levels}{number of levels to retain}

\item{include.other}{if the other factor levels are to be retained}
}
\description{

}
\details{
Function to create an other category in functions where there are too many different levels to present cleanly
}
\author{
Ben Downe
}

